add errorbar stories
